Fixed minor issue with position buffers
Depending on the time of day, I might have today's data for some instruments, but only yesterday's data for others.

When this happens, get_notional_position returns a DataFrame with a nan row for today, which results in pos_buffers also ending with a nan row.  This solves that problem.

It might be better to do this in get_notional_position, but doing it here seemed less likely to have surprising consequences.  In any case, I don't think it does any harm.