Handle unexpected asset classes without failing
IB has a lot of possible values that could be returned here.  

I think it's preferable to just resolve a minimal set of properties like you do for cash, so people with other assets in their IB account don't have to make code changes.

However, I would understand if you prefer an exception to be raised.