Fix BITCOIN expiry in roll config
I recall doing a PR for this change a while ago, but apparently it got lost.

With the ability to import, export, and edit roll config directly in Mongo, it's easy to make a change in one place, but not in another. In fact, I'm not even sure if I should create a PR for this, or an issue.

Related question: is there a good reason for duplicating instrument & roll config (stored in both file and database)?  IB config seems to work fine being read directly from the file.